% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils_batchFillAndWrite.R
\name{batch_fillAndWrite}
\alias{batch_fillAndWrite}
\title{Utility Batch Fill and Write}
\usage{
batch_fillAndWrite(
  data,
  pgConnParam,
  tableName,
  batchSize = 100,
  database = "PostgreSQL"
)
}
\arguments{
\item{data}{A tibble of data}

\item{pgConnParam}{The connection strings required to connect to the PostgreSQL
database of choice.}

\item{tableName}{The table of the PostgreSQL table this data is to be inserted
into. It should include `\\"` around names requiring preservation of case.
Optionally: It can include the SQL Field name specification to deal with a
tibble that may not be ordered the same as the SQL Table. This should look
like: `"\"tableName\" (\"fieldOneName\",\"fieldTwoName\")"`}

\item{batchSize}{How many records to execute at a time.}

\item{database}{What type of database is it going into? Options are: PostgreSQL
or SQLite.}
}
\value{
Returns nothing
}
\description{
This is a utility function used in modSim to take a set of data, break it up
 into batches and write it to the required PostgreSQL database.
}
\note{
Location: ./R/fct_utils_batchFillAndWrite.R
}
\author{
Neil Kester, \email{nkester1@jhu.edu}
}
