% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_step1_createModSimDb.R
\name{createModSimDb}
\alias{createModSimDb}
\title{Create ModSim PostgreSQL Database}
\usage{
createModSimDb(connParamList)
}
\arguments{
\item{connParamList}{A five element named list containing the following elements:
"pgHost", "pgPort", "pgUser", "pgPass", and "pgDb".}
}
\value{
This returns messages as the build progresses and then a `NULL` response.
}
\description{
This function automates the creation of the specific PostgreSQL database used
 to support this thesis work. It takes only a list of connection parameters
 required to connect to the PostgreSQL instance. From that, it creates all
 tables and materialized views used elsewhere in the analytic work flow. This
 is purpose built and not meant to be a general use function.
}
\examples{
createModSimDb(connParamList = pgConnParam)

}
\author{
Neil Kester, \email{nkester1@jhu.edu}
}
