% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_step2_low_etlLosData.R
\name{etlLosData}
\alias{etlLosData}
\title{ETL Line of Sight Data}
\usage{
etlLosData(mongoConnParam, pgConnParam, designPoint)
}
\arguments{
\item{mongoConnParam}{This is a two element named list including the "mongoUri"
which includes the user name and password and a single character string and
the "mongoDb" name as a character string.}

\item{pgConnParam}{A five element named list containing the following elements:
"pgHost", "pgPort", "pgUser", "pgPass", and "pgDb".}

\item{designPoint}{This is a single character string with the designPoint you
would like to extract from the MongoDB and place into the PostgreSQL database.
If multiple designPoints are required then execute this function multiple
times. Note that this pulls ALL iterations executed for that designPoint.}
}
\value{
This returns messages to the console updating the user on the function's
  status but returns no information.
}
\description{
This function queries the LOSTargetStatus collection in the simulation
 MongoDB, transforms and un-nests it and then writes it to the PostgreSQL
 tables  created by the `creatModSimDb` function.
}
\author{
Neil Kester, \email{nkester1@jhu.edu}
}
