% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_step3_high_graphMean.R
\name{graphMean}
\alias{graphMean}
\title{Graph the Mean Data}
\usage{
graphMean(graphData, losColor, acqColor, errorbars = TRUE)
}
\arguments{
\item{graphData}{This is the mean data resulting from `multiDesignPointAndSensorDataPrep`
function}

\item{losColor}{A color string}

\item{acqColor}{A color string}

\item{errorbars}{Boolean}
}
\value{
A ggplot2 plot
}
\description{
Graph the Mean Data
}
\author{
Neil Kester, \email{nkester1@jhu.edu}
}
