% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_step3_high_graphTargetColumnByDesignPoint.R
\name{graphTargetColumnByDesignPoint}
\alias{graphTargetColumnByDesignPoint}
\title{Graph the Sensor - Target Column Chart, Faceted by Design Point}
\usage{
graphTargetColumnByDesignPoint(
  designPoints,
  pgConnParam,
  sensorForce,
  targetForce,
  sensor
)
}
\arguments{
\item{designPoints}{This is a character vector of names describing designPoints
as stored in the PostgreSQL datanase.}

\item{sensorForce}{This is either "BLUEFORCE" or "REDFORCE".}

\item{targetForce}{This is either "BLUEFORCE" or "REDFORCE".}

\item{connParamList}{A five element named list containing the following elements:
"pgHost", "pgPort", "pgUser", "pgPass", and "pgDb".}

\item{sensors}{This is a character vector of sensor entity names you want to
structure data for.}
}
\value{
This returns a list of two elements, the data and the plot it made.
}
\description{
This function accepts parameters, fetches data, transforms it, and produces a
 graph that shows the number of target entities a sensor entity has line of
 sight with and has acquired. This is further broken down by color that
 corresponds to the target's name.
}
\note{
Location: ./R/fct_step3_high_graphTargetColumnByDesignPoint.R

RMarkdown location: ./inst/step3_plotting/Step3_plotting.Rmd
}
