% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoUnnest.R
\name{low_mapAndUnnest}
\alias{low_mapAndUnnest}
\title{Row-wise Map and Unnest JSON by column}
\usage{
low_mapAndUnnest(data, unnestCols)
}
\description{
This function is intended to be used in a row-wise application to map and
 unnest multilevel columns resulting from nested JSON objects.
}
