% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_step3_low_queryData.R
\name{low_queryData}
\alias{low_queryData}
\title{Query ModSim Materialized Views}
\usage{
low_queryData(pgConnParam, sensorForce, targetForce, designPoint)
}
\arguments{
\item{sensorForce}{This is either "BLUEFORCE" or "REDFORCE".}

\item{targetForce}{This is either "BLUEFORCE" or "REDFORCE".}

\item{designPoint}{This is a string name describing the designPoint as stored
in the PostgreSQL datanase.}

\item{connParamList}{A five element named list containing the following elements:
"pgHost", "pgPort", "pgUser", "pgPass", and "pgDb".}
}
\value{
This returns a named list of two elements, the results from the LOS
 and the ACQ Materialized View queries.
}
\description{
This function queries the materialized views created in Step 1 and filled in
 Step 2.
}
\author{
Neil Kester, \email{nkester1@jhu.edu}
}
