% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_step2_low_mapSensorsAndEntities.R
\name{mapSensorsAndEntities}
\alias{mapSensorsAndEntities}
\title{Map Sensors and Entities}
\usage{
mapSensorsAndEntities(mongoUri, mongoDb, mongoCollection, desingPoint)
}
\arguments{
\item{mongoUri}{This is a double quoted character string of the connection
object required for the simulation's MongoDB.}

\item{mongoDb}{This is a double quoted character string of the database name
containing the required files.}

\item{mongoCollection}{This is a double quoted character string of the mongo
collection within the named database containing the required information.}

\item{designPoint}{This is a single character string with the designPoint you
would like to extract from the MongoDB and place into the PostgreSQL database.
If multiple designPoints are required then execute this function multiple
times. Note that this pulls ALL iterations executed for that designPoint.}
}
\value{
This function returns a four element named list of tibbles that
 describe all sensors and entities, their IDs and description.
}
\description{
This function accepts the same inputs as the basic MongoDB connection information
 and submits aggregation pipelines to MongoDB to return information about the
 sensors and entities present in the specified designPoint.
}
\note{
Location: ./R/fct_step2_low_mapSensorsAndEntities.R
}
\author{
Neil Kester, \email{nkester1@jhu.edu}
}
