% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notneeded_fct_utils_mongoUnnest.R
\name{mid_multiColRowwiseUnnest}
\alias{mid_multiColRowwiseUnnest}
\title{Multi Column Row-wise Unnest}
\usage{
mid_multiColRowwiseUnnest(data, unnestCols)
}
\arguments{
\item{unnestCols}{A character vector of column names to apply the `los_low_mapAndUnnest`
function to.}

\item{queryResults}{The results from a mongolite query. It should be a multi-level
data frame.}
}
\value{
This returns a tibble.
}
\description{
Execute the low function `los_low_mapAndUnnest` over many rows and applied
 to multiple columns iteratively. Columns are unnested in the order entered
 in the `cols` character vector.
}
\author{
Neil Kester, \email{nkester1@jhu.edu}
}
