% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoUnnest.R
\name{mongoUnnest}
\alias{mongoUnnest}
\title{Query and/or Transform MongoDB}
\usage{
mongoUnnest(
  mongoUri = "",
  mongoDb = "",
  mongoCollection = "",
  mongoQuery = "{}",
  mongoFields = "{}",
  mongoData = "",
  unnestCols = ""
)
}
\arguments{
\item{mongoUri}{This is a double quoted character string of the connection
object required for the simulation's MongoDB.}

\item{mongoDb}{This is a double quoted character string of the database name
containing the required files.}

\item{mongoCollection}{This is a double quoted character string of the mongo
collection within the named database containing the required information.}

\item{mongoQuery}{This is a double quoted character string containing the json
query to pass to the MongoDB. Note that users should escape nested double
quotes with the \ escape character.}

\item{mongoFields}{This is a double quoted json object of collection fields
that should be returned (annoted with `true`) and those that should not be
returned (annoted with `false`).}

\item{mongoData}{This is a data frame passed to this function to unnest
specific columns. If this is used, do not provide the `mongoUri`, `mongoDb`,
`mongoCollection`, `mongoQuery`, or `mongoFields` pramaters.}

\item{unnestCols}{This is a character vector of the columns you need to unnest.}
}
\value{
This function returns an unnested tibble in which each row is an
 observation of a line of sight event.
}
\description{
This function accepts up to five parameters to describe the connection, query,
 and fields to return. In the process, it conducts specific unnesting functions
 required for the collection in question. As such, it is not appropriate for
 general use.
}
