% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/fct_step3_high_multiDesignPointAndSensorDataPrep.R
\name{multiDesingPointAndSensorDataPrep}
\alias{multiDesingPointAndSensorDataPrep}
\title{Query and Prepare Data for Multiple Design Points and Sensors}
\usage{
multiDesingPointAndSensorDataPrep(
  pgConnParam,
  sensorForce,
  targetForce,
  designPoints,
  sensors
)
}
\arguments{
\item{sensorForce}{This is either "BLUEFORCE" or "REDFORCE".}

\item{targetForce}{This is either "BLUEFORCE" or "REDFORCE".}

\item{designPoints}{This is a character vector of names describing designPoints
as stored in the PostgreSQL datanase.}

\item{sensors}{This is a character vector of sensor entity names you want to
structure data for.}

\item{connParamList}{A five element named list containing the following elements:
"pgHost", "pgPort", "pgUser", "pgPass", and "pgDb".}
}
\value{
This returns a two element named list of consolidated data at the iteration
 and designPoint level.
}
\description{
This function executes and integrates the two Step 3 low functions across a
 number of designPoints and sensor entities.
}
\author{
Neil Kester, \email{nkester1@jhu.edu}
}
