% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mongo_sensorAcquisition.R
\name{sensorAcquisition}
\alias{sensorAcquisition}
\title{Structure Sensor-Target Acquisitions}
\usage{
sensorAcquisition(
  mongoUri,
  mongoDb,
  mongoCollection,
  mongoQuery,
  mongoFields,
  recursiveUnnests
)
}
\arguments{
\item{mongoUri}{This is a double quoted character string of the connection
object required for the simulation's MongoDB.}

\item{mongoDb}{This is a double quoted character string of the database name
containing the required files.}

\item{mongoCollection}{This is a double quoted character string of the mongo
collection within the named database containing the required information.}

\item{mongoQuery}{This is a double quoted character string containing the json
query to pass to the MongoDB. Note that users should escape nested double
quotes with the \ escape character.}

\item{mongoFields}{This is a double quoted json object of collection fields
that should be returned (annoted with `true`) and those that should not be
returned (annoted with `false`).}

\item{recursiveUnnests}{This is a character vector of sequential unnests to
perform. These are executed in the order they are listed.}
}
\value{
This returns a tibble where every row is a sensorID to targetID pair
 and contains the detectionLevel as well as the previousDetectionLevel.
}
\description{
This function queries the MongoDB and un-nests the C2 Simulation Message
 collection. This is unique because all of the nests are recursive meaning
 they must be performed as a loop.
}
\author{
Neil Kester, \email{nkester1@jhu.edu}
}
