% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_step1_createModSimDb.R
\name{Step1_createModSimDb}
\alias{Step1_createModSimDb}
\title{Create ModSim PostgreSQL Database}
\usage{
Step1_createModSimDb(connParamList, Materialized = FALSE)
}
\arguments{
\item{connParamList}{A five element named list containing the following elements:
"pgHost", "pgPort", "pgUser", "pgPass", and "pgDb".}

\item{Materialized}{This is a boolean response if you want to make the view
materialized or not. Materialized views greatly improve the performance of
queries although they duplicate data in memory so they can be costly to
disk space. Default is FALSE.}
}
\value{
This returns messages as the build progresses and then a `NULL` response.
}
\description{
This function automates the creation of the specific PostgreSQL database used
 to support this thesis work. It takes only a list of connection parameters
 required to connect to the PostgreSQL instance. From that, it creates all
 tables and materialized views used elsewhere in the analytic work flow. This
 is purpose built and not meant to be a general use function.
}
\note{
R script location: ./R/fct_step1_createModSimDb.R

RMarkdown location: ./inst/step1_CreatePgDataBase/Step1_configurePostgresDb.Rmd
}
\examples{
createModSimDb(connParamList = pgConnParam)

}
\author{
Neil Kester, \email{nkester1@jhu.edu}
}
