% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_step2_high_queryMongoAndFillPg.R
\name{Step2_queryMongoAndFillPg}
\alias{Step2_queryMongoAndFillPg}
\title{Query MongoDB And Fill PostgreSQL}
\usage{
Step2_queryMongoAndFillPg(
  mongoConnParam,
  pgConnParam,
  designPoint,
  batchSize = 100,
  Materialized = FALSE
)
}
\arguments{
\item{mongoConnParam}{This is a two element named list including the "mongoUri"
which includes the user name and password and a single character string and
the "mongoDb" name as a character string.}

\item{pgConnParam}{A five element named list containing the following elements:
"pgHost", "pgPort", "pgUser", "pgPass", and "pgDb".}

\item{designPoint}{This is a single character string with the designPoint you
would like to extract from the MongoDB and place into the PostgreSQL database.
If multiple designPoints are required then execute this function multiple
times. Note that this pulls ALL iterations executed for that designPoint.}

\item{batchSize}{A numeric integer representing how many records you want to
write to the PostgreSQL database at a time.}

\item{Materialized}{This is a boolean response if you want to make the view
materialized or not. Materialized views greatly improve the performance of
queries although they duplicate data in memory so they can be costly to
disk space. Default is FALSE.}
}
\value{
This returns messages to the console updating the user on the function's
  status but returns no information.
}
\description{
This is the high level function executed in step 2 to extract data from the
  simulation's No-SQL MongoDB, transform it, and load it into the analytic
  PostgreSQL relational database created with `createModSimDb`. This is
  executed for one design point at a time.
}
\note{
Location: ./R/fct_step2_high_queryMongoAndFillPg.R

RMarkdown location: ./inst/step2_queryMongoAndFillPg/Step2_queryMongoAndFillPg.Rmd
}
\author{
Neil Kester, \email{nkester1@jhu.edu}
}
