% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_step3_low_boxPlotOfDifferences.R
\name{Step3_boxPlotOfDifferences}
\alias{Step3_boxPlotOfDifferences}
\title{Calculate and Produce Boxplot of data across iterations.}
\usage{
Step3_boxPlotOfDifferences(iterationData)
}
\arguments{
\item{iterationData}{raw data without summarization}
}
\value{
plot and data
}
\description{
Calculate and Produce Boxplot of data across iterations.
}
\note{
Location: ./R/fct_step3_low_boxPlotOfDifferences.R

RMarkdown location: ./inst/step3_plotting/Step3_plotting.Rmd
}
