% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_step3_low_meanDiffByDesignPointAndTime.R
\name{Step3_meanDiffByDesignPointAndTime}
\alias{Step3_meanDiffByDesignPointAndTime}
\title{Find the mean difference between the # of LOS and ACQ per timestep}
\usage{
Step3_meanDiffByDesignPointAndTime(iterationData)
}
\arguments{
\item{iterationData}{from modSim::Step3_multiDesingPointAndSensorDataPrep}
}
\description{
Find the mean difference between the # of LOS and ACQ per timestep
}
\note{
Location: ./R/fct_step3_low_targetColumnGraphData.R

RMarkdown location: ./inst/step3_plotting/Step3_plotting.Rmd
}
