% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/fct_step3_high_multiDesignPointAndSensorDataPrep.R
\name{Step3_multiDesingPointAndSensorDataPrep}
\alias{Step3_multiDesingPointAndSensorDataPrep}
\title{Query and Prepare Data for Multiple Design Points and Sensors}
\usage{
Step3_multiDesingPointAndSensorDataPrep(
  pgConnParam,
  sensorForce,
  targetForce,
  designPoint,
  sensors
)
}
\arguments{
\item{sensorForce}{This is either "BLUEFORCE" or "REDFORCE".}

\item{targetForce}{This is either "BLUEFORCE" or "REDFORCE".}

\item{sensors}{This is a character vector of sensor entity names you want to
structure data for.}

\item{connParamList}{A five element named list containing the following elements:
"pgHost", "pgPort", "pgUser", "pgPass", and "pgDb".}

\item{designPoints}{This is a character vector of names describing designPoints
as stored in the PostgreSQL datanase.}
}
\value{
This returns a two element named list of consolidated data at the iteration
 and designPoint level. DesignPoint level data is aggregated across the iterations
 executed for each designPoint. Metrics calculated for that aggregation include
 the mean, variance, standard deviation, and standard error.
}
\description{
This function executes and integrates the two Step 3 low functions across a
 number of designPoints and sensor entities.
}
\note{
Location: ./R/fct_step3_high_multiDesignPointAndSensorDataPrep.R

RMarkdown location: ./inst/step3_plotting/Step3_plotting.Rmd
}
\author{
Neil Kester, \email{nkester1@jhu.edu}
}
