% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_step3_low_targetColumnGraphData.R
\name{targetColumnGraphData}
\alias{targetColumnGraphData}
\title{Target-Sensor Entity Data}
\usage{
targetColumnGraphData(losMatViewData, acqMatViewData, sensor)
}
\arguments{
\item{losMatViewData}{A tibble resulting from the `low_queryData` function.}

\item{acqMatViewData}{A tibble resulting from the `low_queryData` function.}

\item{sensor}{A single character string for the sensor you want to structure
data for. For multiple sensors, execute this function multiple times.}
}
\value{
This function returns a single tibble of time, sensor name, iteration,
 design point, target name, and type (LOS or ACQ).
}
\description{
This function prepares the LOS and ACQ data at the target level by time step
 rather than the design point level. This allows us to understand which targets
 the chosen sensor has line of sight with and which it is able to acquire.
}
\note{
Location: ./R/fct_step3_low_targetColumnGraphData.R

RMarkdown location: ./inst/step3_plotting/Step3_plotting.Rmd
}
\author{
Neil Kester, \email{nkester1@jhu.edu}
}
