% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notneeded_fct_utils_mongoUnnest.R
\name{low_mapAndUnnest}
\alias{low_mapAndUnnest}
\title{Row-wise Map and Un-nest JSON by column}
\usage{
low_mapAndUnnest(data, unnestCol)
}
\arguments{
\item{data}{This is a nested row, provided by the `mid_multiColRowwiseUnnest`
function.}

\item{unnestCol}{This should be the single column that must be un-nested.}
}
\description{
This function is intended to be used in a row-wise application to map and
 un-nest multilevel columns resulting from nested JSON objects.
}
\author{
Neil Kester, \email{nkester1@jhu.edu}
}
