% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mongo_mapSensorsAndEntities.R
\name{mapSensorsAndEntities}
\alias{mapSensorsAndEntities}
\title{Map Sensors and Entities}
\usage{
mapSensorsAndEntities(
  mongoUri,
  mongoDb,
  mongoCollection,
  mongoFields,
  mongoQuery
)
}
\arguments{
\item{mongoUri}{This is a double quoted character string of the connection
object required for the simulation's MongoDB.}

\item{mongoDb}{This is a double quoted character string of the database name
containing the required files.}

\item{mongoCollection}{This is a double quoted character string of the mongo
collection within the named database containing the required information.}

\item{mongoFields}{This is a double quoted json object of collection fields
that should be returned (annoted with `true`) and those that should not be
returned (annoted with `false`).}

\item{mongoQuery}{This is a double quoted character string containing the json
query to pass to the MongoDB. Note that users should escape nested double
quotes with the \ escape character.}
}
\value{
This function returns a four element named list of tibbles that
 describe all sensors and entities, their IDs and description.
}
\description{
This function accepts the same inputs as the `modSim::mongoUnnest` function
 minus the `unnestCols` parameter. It executes the requested query which is
 expected to be from the `AcquireModel.state.sensors` collection.
}
\author{
Neil Kester, \email{nkester1@jhu.edu}
}
